requirements.txt
kdl-py==1.2.0
loguru==0.7.3
websockets==15.0
SCPC @ https://github.com/Electro3020/SCPC/archive/refs/heads/main.zip
